// packages/plugin-gated-storage/src/types.ts
import { Content } from "@elizaos/core";
import { type StorageService } from "./services/storage.service.ts";

export interface GateDataConfig {
    provider: {
        table_id: string;
        orbis_gateway: string;
        ceramic_gateway: string;
        private_key: string;
        context_id: string;
    };
}

export interface WeatherData {
    location: string;
    temperature: number;
    humidity: number;
    windSpeed: number;
    description: string;
    units: "metric" | "imperial";
}

export interface GateActionContent extends Content {
    text: string;
}

export interface GateDataProviderResponseGet {
    success: boolean;
    provider?: StorageService;
    error?: string;
}

export interface NonceProviderResponseGet {
    success: boolean;
    nonce?: string;
    error?: string;
}